feat: convert timestamps from seconds to milliseconds
BREAKING CHANGE: All internal timestamps (block times, consensus
parameters) now use milliseconds instead of seconds. This
changes the genesis block, invalidates existing chain data, and requires
full re-synchronization.

Key changes:
- Block header nTime field stores milliseconds
- GetTime() and GetAdjustedTime() return milliseconds
- Consensus parameters (nPowTargetTimespan, nPowTargetSpacing) scaled to ms
- New genesis blocks generated for mainnet, testnet, and regtest
- Updated ISO8601 formatting to support millisecond precision
- All time comparisons and calculations adjusted for ms resolution
- BIP68 sequence locks convert ms block times to 512-second units
- closes #6

Also includes:
- Add -testactivationheight=name@height CLI option replacing -segwitheight
  for configuring buried deployment heights (bip34, dersig, cltv, csv, segwit)
- Add BuriedDeployment enum and deploymentinfo module
- Fix MwebFeeFromWeight overflow protection in feerate.cpp
- Add doc/timestamp-semantics.md documenting unit conventions
- Update functional tests and test data for new timestamp format
- Update MWEB to enable earlier in main and test networks, closes #11

+++ src/wallet/wallet.h
@@ -303,7 +303,7 @@ public:
     mapValue_t mapValue;
     std::vector<std::pair<std::string, std::string> > vOrderForm;
     unsigned int fTimeReceivedIsTxTime;
-    unsigned int nTimeReceived; //!< time received by this node
+    uint64_t nTimeReceived; //!< time received by this node
     /**
      * Stable timestamp that never changes, and reflects the order a transaction
      * was added to the wallet. Timestamp is based on the block time for a
@@ -313,7 +313,7 @@ public:
      * were added to the wallet. More details can be found in
      * CWallet::ComputeTimeSmart().
      */
-    unsigned int nTimeSmart;
+    uint64_t nTimeSmart;
     /**
      * From me flag is set to 1 for transactions that were created by the wallet
      * on this bitcoin node, and set to 0 for transactions that were created
@@ -451,7 +451,7 @@ public:
         }
 
         ReadOrderPos(nOrderPos, mapValue);
-        nTimeSmart = mapValue.count("timesmart") ? (unsigned int)atoi64(mapValue["timesmart"]) : 0;
+        nTimeSmart = mapValue.count("timesmart") ? (uint64_t)atoi64(mapValue["timesmart"]) : 0;
 
         ReadPegoutIndices(pegout_indices, mapValue);
         mweb_wtx_info = mapValue.count("mweb_info") ? boost::make_optional(MWEB::WalletTxInfo::FromHex(mapValue["mweb_info"])) : boost::none;
@@ -1035,7 +1035,7 @@ public:
     bool EncryptWallet(const SecureString& strWalletPassphrase);
 
     void GetKeyBirthTimes(std::map<CKeyID, int64_t> &mapKeyBirth) const EXCLUSIVE_LOCKS_REQUIRED(cs_wallet);
-    unsigned int ComputeTimeSmart(const CWalletTx& wtx) const;
+    uint64_t ComputeTimeSmart(const CWalletTx& wtx) const;
 
     /**
      * Increment the next transaction order id
